fix(characters): i64 EXP, localized status labels, responsive tech tree
- EXP: read as i64 (a high-level player's total EXP exceeds i32::MAX and
  wrapped negative, e.g. -1.7B at Lv80).
- Status points: saves store the localized display name in the server's
  language, so a Japanese server showed raw "最大HP" etc. Map the known enum +
  Japanese forms to canonical English; unknown strings still fall back.
- Tech tree: tiles and icons now scale with viewport width (CSS clamp + a
  CSS-variable-driven TechIcon) so the tree fills wide monitors instead of
  sitting small and centred.

+++ bridge/psm-bridge/src/server.rs
@@ -225,7 +225,7 @@ async fn player_detail(State(state): State<ServerState>, Path(id): Path<String>)
 struct PlayerDetail {
     summary: PlayerSummary,
     level: i32,
-    exp: i32,
+    exp: i64,
     /// Stat-point allocations, e.g. `{ "MaxHP": 3, "MaxSP": 2, ... }`.
     status_points: BTreeMap<String, i32>,
     ext_status_points: BTreeMap<String, i32>,
